Add AiManager and refactor event handling; update entity management interfaces

+++ GameEngine/Util/CoordinateSystem.cs
@@ -55,8 +55,8 @@ public static class CoordinateSystem
 
     public static (byte ChunkX, byte ChunkY, byte LocalX, byte LocalY) WorldToChunkAndLocal(int worldX, int worldY)
     {
-        byte chunkX = worldX / Config.CHUNK_SIZE;
-        byte chunkY = worldY / Config.CHUNK_SIZE;
+        byte chunkX = (byte)(worldX / Config.CHUNK_SIZE);
+        byte chunkY = (byte)(worldY / Config.CHUNK_SIZE);
         byte localX = (byte)(worldX % Config.CHUNK_SIZE);
         byte localY = (byte)(worldY % Config.CHUNK_SIZE);
         return (chunkX, chunkY, localX, localY);
